Added basic client-server interfacing.
Currently, just added a flag to use the server versions of functions if a client flag is on. My echo function works, but nothing with the atomic decorator works for now. However, basic communication works. And the client is able to shut the server down. There's some TODOs that need to be taken care of. Mainly the fact that if the server address is in use then attempt to create a proxy instead of create a full server.