Use an alternate method to get the current bv
If the focus is on the sidebar, then bn_context.binaryView is None.
However, we can get the current view and then ask it for its data, which
gives a valid binary view. I'm not entirely sure if this is the right
way to get this information but it works for now and fixes a bug where
you couldn't configure BinSync if the side panel was open.